python/pyproject.toml
[tool.poetry]
name = "supermarket-receipt"
version = "0.1.0"
description = "Refactoring Kata code"
authors = ["Nicolas Delaby"]
readme = "README.md"
packages = [{include = "supermarket_receipt"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dateutil = "^2.8.2"
jinja2 = "^3.1.2"


[tool.poetry.group.dev.dependencies]
approvaltests = "^8.4.1"
pytest-approvaltests = "^0.2.4"
texttest = "^4.3.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
ruff = "^0.0.278"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["B", "I", "UP", "C4", "ISC", "PT", "RSE", "TRY", "RUF"]

target-version = "py311"
